[Ritesh] | 1. Update package name 2. Added Dao and Service layer for reference 3. Minor fixes
