--- api/src/main/java/org/bahmni/module/ipd/service/impl/ScheduleServiceImpl.java
@@ -1,42 +1,0 @@
-package org.bahmni.module.ipd.service.impl;
-
-import org.bahmni.module.ipd.dao.ScheduleDAO;
-import org.bahmni.module.ipd.model.Schedule;
-import org.bahmni.module.ipd.service.ScheduleService;
-import org.openmrs.api.APIException;
-import org.openmrs.api.impl.BaseOpenmrsService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional
-public class ScheduleServiceImpl extends BaseOpenmrsService implements ScheduleService {
-	
-	private static final Logger log = LoggerFactory.getLogger(ScheduleServiceImpl.class);
-	
-	private final ScheduleDAO scheduleDAO;
-
-	@Autowired
-	public ScheduleServiceImpl(ScheduleDAO scheduleDAO) {
-		this.scheduleDAO = scheduleDAO;
-	}
-
-	@Override
-	@Transactional(readOnly = true)
-	public Schedule getSchedule(Integer scheduleId) throws APIException {
-		return scheduleDAO.getSchedule(scheduleId);
-	}
-	
-	@Override
-	public Schedule saveSchedule(Schedule schedule) throws APIException {
-		return scheduleDAO.saveSchedule(schedule);
-	}
-
-	@Override
-	public void purgeSchedule(Schedule schedule) throws APIException {
-		scheduleDAO.purgeSchedule(schedule);
-	}
-}
