+++ api/src/main/java/org/openmrs/module/ipd/api/IPDActivator.java
@@ -1,0 +1,22 @@
+package org.openmrs.module.ipd.api;
+
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.openmrs.module.BaseModuleActivator;
+
+/**
+ * This class contains the logic that is run every time this module is either started or shutdown
+ */
+public class IPDActivator extends BaseModuleActivator {
+	
+	private final Log log = LogFactory.getLog(this.getClass());
+	
+	public void startup() {
+		log.info("Starting Basic Module");
+	}
+	
+	public void shutdown() {
+		log.info("Shutting down Basic Module");
+	}
+	
+}
