+++ api/src/main/java/org/openmrs/module/ipd/api/dao/ScheduleDAO.java
@@ -1,0 +1,15 @@
+package org.openmrs.module.ipd.api.dao;
+
+import org.openmrs.module.ipd.api.model.Schedule;
+import org.openmrs.api.db.DAOException;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ScheduleDAO {
+	
+	Schedule getSchedule(Integer scheduleId) throws DAOException;
+	
+	Schedule saveSchedule(Schedule schedule) throws DAOException;
+
+	void purgeSchedule(Schedule schedule) throws DAOException;
+}
