+++ api/src/main/java/org/openmrs/module/ipd/api/service/ScheduleService.java
@@ -1,0 +1,19 @@
+package org.openmrs.module.ipd.api.service;
+
+import org.openmrs.module.ipd.api.model.Schedule;
+import org.openmrs.api.APIException;
+import org.openmrs.api.OpenmrsService;
+import org.springframework.stereotype.Service;
+
+@Service
+public interface ScheduleService extends OpenmrsService {
+	
+//	@Authorized({ PrivilegeConstants.EDIT_IPD_SCHEDULES })
+	Schedule getSchedule(Integer scheduleId) throws APIException;
+	
+//	@Authorized({ PrivilegeConstants.EDIT_IPD_SCHEDULES })
+	Schedule saveSchedule(Schedule schedule) throws APIException;
+
+//	@Authorized({ PrivilegeConstants.EDIT_IPD_SCHEDULES })
+	void purgeSchedule(Schedule schedule) throws APIException;
+}
