+++ api/src/main/resources/liquibase.xml
@@ -5,7 +5,7 @@
                    xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
          http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-2.0.xsd">
 
-    <changeSet id="ipd_schedule_202307270752" author="tw">
+    <changeSet id="ipd_schedule_202307270755" author="tw">
         <preConditions onFail="MARK_RAN">
             <not>
                 <tableExists tableName="ipd_schedule"/>
@@ -13,7 +13,7 @@
         </preConditions>
         <createTable tableName="ipd_schedule">
             <column autoIncrement="true" name="schedule_id" type="int"> <constraints nullable="false" primaryKey="true"/> </column>
-            <column name="uuid" type="char(38)" />
+            <column name="uuid" type="varchar(38)" />
             <column name="for_reference_id" type="int"> <constraints nullable="false"/> </column>
             <column name="by_reference_id" type="int"> <constraints nullable="false"/> </column>
             <column name="service_type_id" type="int"> <constraints nullable="true"/> </column>
@@ -44,7 +44,7 @@
                                  referencedTableName="concept" referencedColumnNames="concept_id"/>
     </changeSet>
 
-    <changeSet id="ipd_reference_202307270852" author="tw">
+    <changeSet id="ipd_reference_202307270855" author="tw">
         <preConditions onFail="MARK_RAN">
             <not>
                 <tableExists tableName="ipd_reference"/>
@@ -54,8 +54,7 @@
         <createTable tableName="ipd_reference">
             <column name="reference_id" type="int" autoIncrement="true"><constraints primaryKey="true"/></column>
             <column name="target_type" type="varchar(255)"><constraints nullable="false"/></column>
-            <column name="target_uuid" type="char(32)"><constraints unique="true"/></column>
-            <column name="reference" type="varchar(255)"><constraints nullable="false"/></column>
+            <column name="target_uuid" type="varchar(38)"><constraints unique="true"/></column>
             <column name="description" type="varchar(255)"/>
             <column name="creator" type="int"><constraints nullable="false"/></column>
             <column name="date_created" type="datetime"><constraints nullable="false"/></column>
@@ -77,12 +76,12 @@
         <addForeignKeyConstraint baseTableName="ipd_reference" baseColumnNames="retired_by"
                                  constraintName="ipd_reference_retired_by_fk"
                                  referencedTableName="users" referencedColumnNames="user_id"/>
-        <addForeignKeyConstraint baseTableName="ipd_schedule" baseColumnNames="for_reference_id"
-                                 constraintName="ipd_schedule_managed_by_reference_fk"
+        <addForeignKeyConstraint baseTableName="ipd_reference" baseColumnNames="for_reference_id"
+                                 constraintName="ipd_reference_managed_by_reference_fk"
                                  referencedTableName="ipd_reference"
                                  referencedColumnNames="reference_id"/>
-        <addForeignKeyConstraint baseTableName="ipd_schedule" baseColumnNames="by_reference_id"
-                                 constraintName="ipd_schedule_for_reference_fk"
+        <addForeignKeyConstraint baseTableName="ipd_reference" baseColumnNames="by_reference_id"
+                                 constraintName="ipd_reference_for_reference_fk"
                                  referencedTableName="ipd_reference"
                                  referencedColumnNames="reference_id"/>
     </changeSet>
