--- omod/src/main/java/org/bahmni/module/ipd/ModuleDescriptor.java
@@ -1,3 +1,0 @@
-package org.bahmni.module.ipd;
-
-public class ModuleDescriptor {}
