--- omod/src/main/java/org/bahmni/module/ipd/service/ScheduleMedicationService.java
@@ -1,8 +1,0 @@
-package org.bahmni.module.ipd.service;
-
-import org.bahmni.module.ipd.contract.ScheduleMedicationRequest;
-import org.bahmni.module.ipd.model.Schedule;
-
-public interface ScheduleMedicationService {
-    Schedule createSchedule(ScheduleMedicationRequest scheduleMedicationRequest);
-}
