+++ omod/src/main/java/org/openmrs/module/ipd/ModuleDescriptor.java
@@ -1,0 +1,3 @@
+package org.openmrs.module.ipd;
+
+public class ModuleDescriptor {}
