+++ omod/src/main/java/org/openmrs/module/ipd/contract/ScheduleMedicationResponse.java
@@ -1,0 +1,34 @@
+package org.openmrs.module.ipd.contract;
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.openmrs.module.ipd.api.model.Schedule;
+import org.openmrs.Concept;
+
+import java.util.Date;
+
+@Builder
+@Getter
+@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+@AllArgsConstructor
+@NoArgsConstructor
+public class ScheduleMedicationResponse {
+    private Integer id;
+    private String patientUuid;
+    private String comments;
+    private Date startDate;
+    private Date endDate;
+
+    public static ScheduleMedicationResponse constructFrom(Schedule schedule) {
+        return ScheduleMedicationResponse.builder()
+            .id(schedule.getId())
+            .patientUuid(schedule.getForReference().getUuid())
+            .comments(schedule.getComments())
+            .startDate(schedule.getStartDate())
+            .endDate(schedule.getEndDate())
+            .build();
+    }
+}
