+++ omod/src/main/java/org/openmrs/module/ipd/service/ScheduleMedicationService.java
@@ -1,0 +1,8 @@
+package org.openmrs.module.ipd.service;
+
+import org.openmrs.module.ipd.contract.ScheduleMedicationRequest;
+import org.openmrs.module.ipd.api.model.Schedule;
+
+public interface ScheduleMedicationService {
+    Schedule createSchedule(ScheduleMedicationRequest scheduleMedicationRequest);
+}
