+++ omod/src/main/resources/config.xml
@@ -14,6 +14,6 @@
 		<require_module>org.bahmni.module.bahmnicore</require_module>
 	</require_modules>
 
-	<activator>org.bahmni.module.ipd.IPDActivator</activator>
+	<activator>org.openmrs.module.ipd.api.IPDActivator</activator>
 </module>
