--- omod/src/test/java/org/bahmni/module/ipd/BaseIntegrationTest.java
@@ -1,7 +1,0 @@
-package org.bahmni.module.ipd;
-
-import org.springframework.test.context.ContextConfiguration;
-
-@ContextConfiguration(locations = {"classpath:TestingApplicationContext.xml"})
-public abstract class BaseIntegrationTest extends BaseWebControllerTest {
-}
