+++ omod/src/test/java/org/openmrs/module/ipd/BaseIntegrationTest.java
@@ -1,0 +1,7 @@
+package org.openmrs.module.ipd;
+
+import org.springframework.test.context.ContextConfiguration;
+
+@ContextConfiguration(locations = {"classpath:TestingApplicationContext.xml"})
+public abstract class BaseIntegrationTest extends BaseWebControllerTest {
+}
